UltimateCube blight buff

+++ EatYourBeetSVG/src/main/java/eatyourbeets/utilities/GameUtilities.java
@@ -547,7 +547,7 @@ public class GameUtilities
         return creature.currentHealth / (float) creature.maxHealth;
     }
 
-    public static float GetActualHealth(AbstractCreature creature)
+    public static int GetActualHealth(AbstractCreature creature)
     {
         return creature.currentHealth + TempHPField.tempHp.get(creature);
     }
